// src/types/film.ts
export type FilmType = {
    id: number;
    Actors: string;
    Awards: string;
    BoxOffice: number;
    Country: string;
    DVD: string;
    Director: string;
    Genre: string;
    Writer: string;
    Runtime: number;
    Plot: string;
    imdbID: string;
    imdbRating: number;
    officialSite: string;
    Year: number;
    Released: string;
    Title: string;
    Poster: string;
}